permit ActiveRecord 4 in gemspec
Previously we could not allow ActiveRecord 4 because the
activerecord-nulldb-adapter gem did not support ActiveSupport 4 in any
released version.

The activerecord-nulldb-adapter maintainers have recently released
version 0.3.0, and this version is compatible with ActiveSupport 4.

Bump the bogus gemspec ActiveRecord version to reflect the fact that we
can now support ActiveRecord 4 as well.