avoid dependency on Coveralls::SimpleCov::Formatter
In 5b8bb36e3e27d4db13541f294a7c42d74eb417b2, bogus gained the ability to
run the test suite without Coveralls. There was a bug in this
implementation: if we do not have coveralls installed, we can't refer to
Coveralls constants later on.

Move the Coverall reference inside the conditional.